Complete relation plotting interactive
Incorrect points are now removed immediately. Points are locked to prevent dragging after placement. A little burst animation plays when points are placed; this is partly for fun, but also to emphasize the that action is happening to touchscreen users whose fingers may be covering the point upon its placement.